# gears/system/resource-group/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Resource Group API
  version: 0.1.0
  description: |
    Resource Group API — a generic hierarchy and membership engine.

    **This document is generated.** It is the resource-group subset of `docs/api/api.json`, which
    `make openapi` produces from the running server. Do not edit it by hand — edit
    `docs/openapi.preamble.yaml` (this preamble) or the code, then run `make openapi-gears`.

    ## Path prefix

    The paths below are gear-relative, exactly as the gear registers them. A deployment may nest
    them under a gateway prefix configured as `api-gateway.prefix_path`; in `config/quickstart.yaml`
    that prefix is `/cf`, and in the local e2e stack there is none. The `servers` entry reflects the
    configuration the document was generated from.

    ## Types API

    `/types-registry/v1/types` is a structured registry, not a raw schema store. A type is created
    from a flat DTO — `code`, `can_be_root`, `allowed_parent_types`, `allowed_membership_types`,
    `metadata_schema` — and is addressed by its `code` (a GTS type path). `can_be_root` has no
    column of its own: it is persisted inside `metadata_schema` under the reserved key
    `__can_be_root`, and reserved `__`-prefixed keys are stripped on read.

    Registration order matters: types referenced in `allowed_parent_types` or
    `allowed_membership_types` must already exist, so base types come before chained ones. Seeding
    is strictly sequential for this reason.

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /resource-group/v1/memberships:
    get:
      description: Retrieve a paginated list of memberships with OData filtering on group_id, resource_type,
        resource_id
      operationId: resource_group.list_memberships
      parameters:
      - description: Maximum number of memberships to return
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          type: string
      - description: |-
          OData v4 filter expression
          - group_id: eq|ne|in
          - resource_type: eq|ne|contains|startswith|endswith|in
          - resource_id: eq|ne|contains|startswith|endswith|in
        in: query
        name: $filter
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_MembershipDto'
          description: Paginated list of memberships
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List memberships
      tags:
      - Resource Group Memberships
      x-odata-filter:
        allowedFields:
          group_id:
          - eq
          - ne
          - in
          resource_id:
          - eq
          - ne
          - contains
          - startswith
          - endswith
          - in
          resource_type:
          - eq
          - ne
          - contains
          - startswith
          - endswith
          - in
  /resource-group/v1/memberships/{group_id}/{resource_type}/{resource_id}:
    delete:
      description: Remove a membership link between a resource group and a resource
      operationId: resource_group.remove_membership
      parameters:
      - description: Group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      - description: GTS type path of the resource type
        in: path
        name: resource_type
        required: true
        schema:
          type: string
      - description: Resource identifier
        in: path
        name: resource_id
        required: true
        schema:
          type: string
      responses:
        '204':
          content:
            application/json:
              schema:
                type: object
          description: Membership removed successfully
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Remove membership
      tags:
      - Resource Group Memberships
    post:
      description: Add a membership link between a resource group and a resource
      operationId: resource_group.add_membership
      parameters:
      - description: Group UUID
        in: path
        name: group_id
        required: true
        schema:
          type: string
      - description: GTS type path of the resource type
        in: path
        name: resource_type
        required: true
        schema:
          type: string
      - description: Resource identifier
        in: path
        name: resource_id
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipDto'
          description: Membership created
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Add membership
      tags:
      - Resource Group Memberships
  /types-registry/v1/types:
    get:
      description: Retrieve a list of GTS resource group type definitions with OData filtering
      operationId: resource_group.list_types
      parameters:
      - description: Maximum number of types to return
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          type: string
      - description: |-
          OData v4 filter expression
          - code: eq|ne|contains|startswith|endswith|in
        in: query
        name: $filter
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_TypeDto'
          description: List of GTS types
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List GTS types
      tags:
      - Resource Group Types
      x-odata-filter:
        allowedFields:
          code:
          - eq
          - ne
          - contains
          - startswith
          - endswith
          - in
    post:
      description: Create a new GTS resource group type definition
      operationId: resource_group.create_type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTypeDto'
        description: Type creation data
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeDto'
          description: Created type
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create a new GTS type
      tags:
      - Resource Group Types
  /types-registry/v1/types/{code}:
    delete:
      description: Delete a GTS resource group type definition
      operationId: resource_group.delete_type
      parameters:
      - description: GTS type path
        in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        '204':
          content:
            application/json:
              schema:
                type: object
          description: Type deleted successfully
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Delete GTS type
      tags:
      - Resource Group Types
    get:
      description: Retrieve a specific GTS type definition by its GTS type path
      operationId: resource_group.get_type
      parameters:
      - description: GTS type path
        in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeDto'
          description: Type found
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get GTS type by code
      tags:
      - Resource Group Types
    put:
      description: Update a GTS resource group type definition (full replacement)
      operationId: resource_group.update_type
      parameters:
      - description: GTS type path
        in: path
        name: code
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTypeDto'
        description: Type update data
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeDto'
          description: Updated type
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Update GTS type
      tags:
      - Resource Group Types
components:
  schemas:
...
    CreateTypeDto:
      description: REST DTO for creating a new GTS type.
      properties:
        allowed_membership_types:
          description: GTS type paths of allowed membership resource types.
          items:
            type: string
          type: array
        allowed_parent_types:
          description: GTS type paths of allowed parent types.
          items:
            type: string
          type: array
        can_be_root:
          description: Whether groups of this type can be root nodes.
          type: boolean
        code:
          description: |-
            GTS type path. Must have prefix `gts.cf.core.rg.type.v1~`.

            Whether the type creates a new tenant scope is derived from the code:
            any path starting with the tenant RG type prefix is a tenant type.
          type: string
        metadata_schema:
          description: Optional JSON Schema for instance metadata.
      required:
      - code
      - can_be_root
      type: object
...
    MembershipDto:
      description: |-
        REST DTO for membership representation.

        Membership responses do NOT include `tenant_id` (derived from group).
      properties:
        group_id:
          description: Group identifier.
          format: uuid
          type: string
        resource_id:
          description: Resource identifier.
          type: string
        resource_type:
          description: GTS type path of the resource type.
          type: string
      required:
      - group_id
      - resource_type
      - resource_id
      type: object
...
    PageInfo:
      properties:
        limit:
          format: int64
          minimum: 0
          type: integer
        next_cursor:
          type:
          - string
          - 'null'
        prev_cursor:
          type:
          - string
          - 'null'
      required:
      - limit
      type: object
...
    Page_MembershipDto:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MembershipDto'
          type: array
        page_info:
          $ref: '#/components/schemas/PageInfo'
      required:
      - items
      - page_info
      type: object
    Page_TypeDto:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TypeDto'
          type: array
        page_info:
          $ref: '#/components/schemas/PageInfo'
      required:
      - items
      - page_info
      type: object
    Problem:
      description: RFC 9457 problem+json. `context` varies by error category.
      properties:
        context:
          type: object
        detail:
          type: string
        instance:
          type: string
        status:
          format: int32
          type: integer
        title:
          type: string
        trace_id:
          type: string
        type:
          type: string
      required:
      - type
      - title
      - status
      - detail
      - context
      type: object
    TypeDto:
      description: REST DTO for GTS type representation.
      properties:
        allowed_membership_types:
          description: GTS type paths of allowed membership resource types
          items:
            type: string
          type: array
        allowed_parent_types:
          description: GTS type paths of allowed parent types
          items:
            type: string
          type: array
        can_be_root:
          description: Whether groups of this type can be root nodes
          type: boolean
        code:
          description: GTS type path
          type: string
        metadata_schema:
          description: Optional JSON Schema for instance metadata
      required:
      - code
      - can_be_root
      - allowed_parent_types
      - allowed_membership_types
      type: object
...
    UpdateTypeDto:
      additionalProperties: false
      description: |-
        REST DTO for updating a GTS type (full replacement via PUT).

        The type registry is a document resource: a type definition is replaced as
        a whole, so strict full replacement is the correct shape for it (baseline
        rule B1.1). Every replaceable field is therefore **required** and an
        omitted key is an error, not "preserve previous value" — `metadata_schema`
        must be sent explicitly as `null` to clear it.

        `deny_unknown_fields` locks the wire envelope, so a client that sends
        `code` (the type's identity, taken from the path and immutable) gets an
        explicit 400 instead of a silently-dropped field.
      properties:
        allowed_membership_types:
          description: GTS type paths of allowed membership resource types.
          items:
            type: string
          type: array
        allowed_parent_types:
          description: GTS type paths of allowed parent types.
          items:
            type: string
          type: array
        can_be_root:
          description: Whether groups of this type can be root nodes.
          type: boolean
        metadata_schema:
          description: |-
            JSON Schema for instance metadata. Required: send `null` to clear it.
            An omitted key is rejected — it used to silently erase the stored
            schema (`type_service.rs` writes the column unconditionally).
      required:
      - can_be_root
      - allowed_parent_types
      - allowed_membership_types
      - metadata_schema
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http